cortex-chatbot/requirements.txt
fastapi
snowflake-connector-python
snowflake
snowflake-ml-python
python-dotenv
magentic
sse-starlette